[wip] `tw_encoding` FFI separation with other public functions
* Move FFI tests to `tw_encoding/tests/ffi`
