+++ rust/tw_encoding/src/lib.rs
@@ -7,4 +7,5 @@
 pub mod base32;
 pub mod base58;
 pub mod base64;
+pub mod ffi;
 pub mod hex;
