+++ rust/tw_encoding/tests/ffi.rs
@@ -1,0 +1,8 @@
+#[path = "ffi/base32_tests.rs"]
+mod base32_tests;
+#[path = "ffi/base58_tests.rs"]
+mod base58_tests;
+#[path = "ffi/base64_tests.rs"]
+mod base64_tests;
+#[path = "ffi/hex_tests.rs"]
+mod hex;
