+++ rust/tw_memory/src/lib.rs
@@ -6,6 +6,7 @@
 
 use std::ffi::{c_char, CString};
 
+/// TODO declare `free_cbyte_array`, `is_cbyte_array_null` FFI.
 #[repr(C)]
 pub struct CByteArray {
     pub data: *mut u8,
@@ -15,10 +16,19 @@ pub struct CByteArray {
 impl From<Vec<u8>> for CByteArray {
     fn from(value: Vec<u8>) -> Self {
         let size = value.len();
-        let mut mut_vec = value.to_vec();
+        let mut mut_vec = value.clone();
         let ptr = mut_vec.as_mut_ptr();
         std::mem::forget(mut_vec);
         CByteArray { data: ptr, size }
+    }
+}
+
+impl CByteArray {
+    pub fn null() -> CByteArray {
+        CByteArray {
+            data: std::ptr::null_mut(),
+            size: 0,
+        }
     }
 }
